+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.1.1 - 2020-09-05
+
 * Allow index queries.
 * Allow field queries with special characters.
 
@@ -10,3 +12,6 @@
 ### Added
 
 * Added `DeserializeQuery` derive macro. Currently it handles only map accesses.
+
+[0.1.1]: https://github.com/pandaman64/serde-query/compare/v0.1.0...v0.1.1
+[0.1.0]: https://github.com/pandaman64/serde-query/releases/tag/v0.1.0